requirements.txt
aniso8601==8.0.0
attrs==19.3.0
click==7.1.1
dnspython==1.16.0
eventlet==0.25.1
Flask==1.1.1
flask-restx==0.2.0
Flask-SocketIO==4.2.1
greenlet==0.4.15
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
MarkupSafe==1.1.1
monotonic==1.5
pyrsistent==0.16.0
python-dotenv==0.12.0
python-engineio==3.12.1
python-socketio==4.5.1
pytz==2019.3
six==1.14.0
Werkzeug==1.0.0
zipp==3.1.0
